Fixed QueryUsersRequest limit & offset + added client and users code samples + added users query integration test

+++ Assets/Plugins/StreamChat/Core/DTO/Models/SortParamDTO.cs
@@ -17,7 +17,7 @@ namespace StreamChat.Core.DTO.Models
     internal partial class SortParamDTO
     {
         [Newtonsoft.Json.JsonProperty("direction", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
-        public double? Direction { get; set; }
+        public int? Direction { get; set; }
 
         [Newtonsoft.Json.JsonProperty("field", Required = Newtonsoft.Json.Required.Default, NullValueHandling = Newtonsoft.Json.NullValueHandling.Ignore)]
         public string Field { get; set; }
